requirements.txt
requests~=2.23.0
requests_toolbelt
urllib3~=1.25.11
cryptography
pyOpenSSL
certifi
idna
pillow
ddddocr
